gcp_model_export/requirements.txt
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
google-cloud== 0.34.0
db-dtypes==1.4.2